nb_requirements.txt
google-genai
pillow
exifread
google-api-core
